yaml utils and job_config changes
