+++ templates/Dockerfile
@@ -19,6 +19,8 @@ COPY . /app
 # ##ENV
 ARG YML_FILE='workflow.yml'
 ENV YML_FILE=$YML_FILE
+ARG JOB_CONFIG
+ENV JOB_CONFIG=$JOB_CONFIG
 ARG JOB_ID
 ENV JOB_ID=$JOB_ID
 ARG AWS_ACCESS_KEY_ID
